Optimize and harden audio engine DSP paths
This commit applies several performance optimizations and architectural
hardenings to the Mixxx audio engine:

1. Performance Optimization: Replaces expensive floating-point divisions
   with multiplications by pre-calculated reciprocals in hot DSP loops
   (SampleUtil, EngineVuMeter).
2. Loop Refactoring: Optimizes linear crossfade routines in SampleUtil
   by using incremental additions instead of per-loop index
   multiplications, reducing CPU cycles in the mixing path.
3. Numeric Hygiene: Standardizes on std:: math functions and ensures
   the use of float literals (e.g., 1.0f) and explicit casts to
   CSAMPLE/float to prevent unintentional double-precision promotions
   in real-time threads.
4. Compiler Hints: Improves const-correctness and restricts pointers in
   critical loops (EngineMixer) to facilitate better auto-vectorization.
5. Code Hardening: Ensures consistent use of namespaced math functions
   (std::pow, std::ceil) to guarantee correct overload selection.

Verified with SampleUtil unit tests and code review.

+++ src/engine/enginexfader.cpp
@@ -10,7 +10,7 @@ const double EngineXfader::kTransformMin = 0.6;
 
 double EngineXfader::getPowerCalibration(double transform) {
     // get the transform_root of -3db (.5)
-    return pow(0.5, 1.0 / transform);
+    return std::pow(0.5, 1.0 / transform);
 }
 
 void EngineXfader::getXfadeGains(double xfadePosition,
@@ -50,14 +50,14 @@ void EngineXfader::getXfadeGains(double xfadePosition,
         }
     } else {
         if (xfadePositionLeft < 0) { // on left side
-            xfadePositionLeft *= -1;
-            *gain2 = 1.0f - static_cast<float>(std::pow(static_cast<float>(xfadePositionLeft), static_cast<float>(transform)));
+            xfadePositionLeft *= -1.0;
+            *gain2 = 1.0f - std::pow(static_cast<float>(xfadePositionLeft), static_cast<float>(transform));
         } else {
             *gain2 = 1.0f;
         }
 
         if (xfadePositionRight > 0) { // right side
-            *gain1 = 1.0f - static_cast<float>(std::pow(static_cast<float>(xfadePositionRight), static_cast<float>(transform)));
+            *gain1 = 1.0f - std::pow(static_cast<float>(xfadePositionRight), static_cast<float>(transform));
         } else {
             *gain1 = 1.0f;
         }
